[Enhancement] Refactor Cleanup Task Handler (#516)
* refactor cleanup task handler

* format

* make base class abstract

* make cleanup task record

* simplify logger

* update test after merge

* update task handler register after merge

* fix test error after merge

* refine call context and ut after merge
