+++ service/common/src/main/java/org/apache/polaris/service/task/TaskExecutorImpl.java
@@ -63,6 +63,9 @@ public class TaskExecutorImpl implements TaskExecutor {
     addTaskHandler(
         new ManifestFileCleanupTaskHandler(
             fileIOSupplier, Executors.newVirtualThreadPerTaskExecutor()));
+    addTaskHandler(
+        new BatchFileCleanupTaskHandler(
+            fileIOSupplier, Executors.newVirtualThreadPerTaskExecutor()));
   }
 
   /**
